requirement.txt
alembic==1.4.1
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
importlib-metadata==1.5.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
zipp==3.1.0
boto3
requests
numpy
asyncio
filemagic
pymysql
pillow
pickle-mixin
gevent
scikit-learn
torch
redis
multiexit
apscheduler==2.1.2
psycopg2-binary